fix daemon mode
src/wdm/resource.c sets the debug value to 1 by default. It used to be 0 by default. This causes a problem in that a value of 1 prevents wdm from running as a daemon.  Line 140 of src/wdm/dm.c contains an if statement that only calls BecomeDaemon if debugLevel == 0. Reported by Noah Meyerhans <noahm@debian.org>